+++ test/app/components/instrumentation_component.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class InstrumentationComponent < ViewComponent::Base
+  include ViewComponent::Instrumentation
+end